requirements.txt
numpy
scipy
appnope==0.1.0
attrdict==2.0.1
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cytoolz==0.9.0.1
decorator==4.3.2
Flask==1.0.2
hexbytes==0.1.0
idna==2.8
ipython==7.3.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
lru-dict==1.1.6
MarkupSafe==1.1.1
parsimonious==0.8.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycryptodome==3.7.3
Pygments==2.3.1
python-decouple==3.1
python-dotenv==0.10.1
redis==3.2.1
requests==2.21.0
rlp==1.1.0
six==1.12.0
toolz==0.9.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
web3==4.8.3
websockets==6.0
Werkzeug==0.14.1
pymongo==3.7.2
